copied 'source ourder to vignette
In previous versions of this code everything was scripted. This moves the final unscripted code for segmentation to a vignette.

There is remaining script in the QA/QC process all included in data-raw/ but as those changes should be reflected in changes to the databases it is consided low priority to clean up that code.